requirements.txt
oci
numpy==1.26.4
pillow
imgaug
requests
tensorflow